src/components/FormComponents/Egenmelding/egenmeldingsdagerFieldUtils.ts
import { add, isBefore, sub } from 'date-fns'

export function currentPeriodDatePicker(
    previous: {
        earliestPossibleDate: Date
        earliestSelectedDate: Date | null
    },
    previousSykmeldingTom: Date | null,
): [earliest: Date, latest: Date] {
    const latest = sub(previous.earliestPossibleDate, { days: 1 })
    const earliest = previous.earliestSelectedDate
        ? sub(previous.earliestSelectedDate, { days: 16 })
        : sub(previous.earliestPossibleDate, { days: 16 })

    // Earliest should be limited by previous sykmelding
    if (previousSykmeldingTom && isBefore(earliest, previousSykmeldingTom)) {
        return [add(previousSykmeldingTom, { days: 1 }), latest]
    } else {
        return [earliest, latest]
    }
}
